Historical Simulation

+++ src/main/java/com/adrian/ng/RiskMeasure.java
@@ -1,7 +1,9 @@
 package com.adrian.ng;
 
+import java.math.BigDecimal;
+
 public interface RiskMeasure {
 
-    double getVar();
+    BigDecimal getVar();
 
 }
